+++ example/javascript/monitor.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const config = require('./config');
-const { monitor } = require('../'); // replace with require('redis-smq');
+const { monitor } = require('../../index'); // replace with require('redis-smq');
 
 monitor(config).listen(() => {
     console.log('It works!');