+++ example/typescript/monitor.ts
@@ -1,0 +1,6 @@
+import { config } from "./config";
+import RedisSMQ from "../../";
+
+RedisSMQ.monitor(config).listen(() => {
+    console.log('It works!');
+});