+++ types/instance.ts
@@ -1,0 +1,5 @@
+export interface InstanceInterface {
+    run(): void,
+    shutdown(): void,
+    isRunning(): boolean,
+}